Better native error handling
- integrated features of evalErrorLocator
